admin회원가입수정

+++ src/main/java/com/kosa/ShareTour/repository/AdminRepository.java
@@ -3,7 +3,7 @@ package com.kosa.ShareTour.repository;
 import com.kosa.ShareTour.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
 
     Admin findByName(String name);
     void deleteByName(String name);
